マップ分岐完成、あとは画像やテキスト変えたり #33
