+++ test3-prototype-April/js/lib/TextAudio.json.js
@@ -65,4 +65,26 @@ export default {
         //     "file":"7a1e50428682eba3.mp3"
         // }
     ],
+    "A2-2":[
+        // {
+        //     "start":0,
+        //     "end":3,
+        //     "file":"7a1e50428682eba3.mp3"
+        // },
+        {
+            "start":4,
+            "end":9,
+            "file":"6fa5803adee16db6.mp3"
+        },
+        // {
+        //     "start":10,
+        //     "end":18,
+        //     "file":"7a1e50428682eba3.mp3"
+        // },
+        // {
+        //     "start":22,
+        //     "end":30,
+        //     "file":"7a1e50428682eba3.mp3"
+        // }
+    ],
 }
