+++ test3-prototype-April/js/main.js
@@ -1,19 +1,22 @@
 "use strict"
-import ScenarioPlayer from './lib/ScenarioPlayer.js';
-// import TextData from './lib/TextDataA.json.js';
-// import TextData from './lib/TextDataZeroDay.json.js';
-import TextData from './lib/scenario_data.json' assert { type: "json" };
+import ScenarioPlayer from './lib/ScenarioPlayer.js'
+// import TextData from './lib/TextDataA.json.js'
+// import TextData from './lib/TextDataZeroDay.json.js'
+import TextData from './lib/scenario_data.json' assert { type: "json" }
 import FrameSizing from "./lib/FrameSizing.js"
-import mapItems from "./lib/mapItems.json.js"
+// import mapItems from "./lib/mapItems.json.js" 
+import mapItemsJson from "./lib/mapItems.json" assert { type: "json" }
+import { CreateMap } from './lib/map.js'
 
 (function(){
     const gameState = {
         textEventId: -1, // いらない
-        nowPart: null,
+        nowPart: undefined,
         FloatCheck: true,
         nowDate: 0,
         eventState: 'title',
         autoPlayingFlag: false, // いらない
+        TextPlayer: undefined
     }
     let TextPlayer
     const TextPlayerList = []
@@ -29,105 +32,109 @@ import mapItems from "./lib/mapItems.json.js"
         // 伝搬チェック用
         // document.querySelectorAll('*').forEach(element => {
         //     element.addEventListener('click',e=>{
-        //         console.log(e.target);
+        //         console.log(e.target)
         //     })
         // });
 
-        // マップアイテムの生成
-        gameState.eventState = 'map'
-        const eleFragment = document.createDocumentFragment()
-        for (const itemId in mapItems) {
-            if (Object.hasOwnProperty.call(mapItems, itemId)) {
-                const item = mapItems[itemId]
-                const itemEle = document.createElement('div')
-                itemEle.id = itemId
-                itemEle.className = 'map-touch'
-                itemEle.dataset.place = item.place
+        CreateMap(gameState)
 
-                // イベント付与
-                itemEle.addEventListener('click',(e) => {
+        // // マップアイテムの生成
+        // gameState.eventState = 'map'
+        // const eleFragment = document.createDocumentFragment()
+        // // シナリオ格納
+        // const mapItems = mapItemsJson[gameState.nowPart ?? 'A0']
+        // for (const itemId in mapItems) {
+        //     if (Object.hasOwnProperty.call(mapItems, itemId)) {
+        //         const item = mapItems[itemId]
+        //         const itemEle = document.createElement('div')
+        //         itemEle.id = itemId
+        //         itemEle.className = 'map-touch'
+        //         itemEle.dataset.place = item.place
+
+        //         // イベント付与
+        //         itemEle.addEventListener('click',(e) => {
                 
-                    const float = document.getElementById('mapWrapper')
-                    float.classList.remove('none')
-                    const pname = document.getElementById('placeName')
-                    const placeName = e.target.dataset.place
-                    pname.innerHTML = placeName
+        //             const float = document.getElementById('mapWrapper')
+        //             float.classList.remove('none')
+        //             const pname = document.getElementById('placeName')
+        //             const placeName = e.target.dataset.place
+        //             pname.innerHTML = placeName
 
-                    const partKey = item.partKey
+        //             const partKey = item.partKey
     
-                    const yesBtn = document.getElementById('mapSelectYes')
-                    const noBtn = document.getElementById('mapSelectNo')
+        //             const yesBtn = document.getElementById('mapSelectYes')
+        //             const noBtn = document.getElementById('mapSelectNo')
     
-                    /**
-                     * 付与するYES/NO選択イベント
-                     * @param {*} e event
-                     */
-                    const selectEve = async(e) => {
-                        float.classList.add('none')
-                        e.target.removeEventListener('click',selectEve,false)//クリックされたボタンのイベントを削除
-                        e.target === yesBtn ? noBtn : yesBtn .removeEventListener('click',selectEve,false)//クリックされなかったボタンのイベントを削除
+        //             /**
+        //              * 付与するYES/NO選択イベント
+        //              * @param {*} e event
+        //              */
+        //             const selectEve = async(e) => {
+        //                 float.classList.add('none')
+        //                 e.target.removeEventListener('click',selectEve,false)//クリックされたボタンのイベントを削除
+        //                 e.target === yesBtn ? noBtn : yesBtn .removeEventListener('click',selectEve,false)//クリックされなかったボタンのイベントを削除
                         
-                        const timer = (s) => {
-                            return new Promise((resolve,reject)=>{
-                                const timerId = setTimeout(() => {
-                                    // clearTimeout(timerId)
-                                    resolve()
-                                }, s);
-                            })
-                        }
+        //                 const timer = (s) => {
+        //                     return new Promise((resolve,reject)=>{
+        //                         const timerId = setTimeout(() => {
+        //                             // clearTimeout(timerId)
+        //                             resolve()
+        //                         }, s);
+        //                     })
+        //                 }
                         
-                        if( e.target === yesBtn ) { // yesが押された
+        //                 if( e.target === yesBtn ) { // yesが押された
 
-                            // 暗転
-                            document.getElementById('darkening-floor').classList.remove('op0')
-                            // タイマー
-                            await timer(1000);
-                            // シナリオ画面へ遷移
-                            document.getElementById('textScreen').classList.remove('none')
-                            document.getElementById('mapScreen').classList.add('none')
-                            // タイマー
-                            await timer(1000);
-                            // 暗転解除
-                            document.getElementById('darkening-floor').classList.add('op0')
+        //                     // 暗転
+        //                     document.getElementById('darkening-floor').classList.remove('op0')
+        //                     // タイマー
+        //                     await timer(1000);
+        //                     // シナリオ画面へ遷移
+        //                     document.getElementById('textScreen').classList.remove('none')
+        //                     document.getElementById('mapScreen').classList.add('none')
+        //                     // タイマー
+        //                     await timer(1000);
+        //                     // 暗転解除
+        //                     document.getElementById('darkening-floor').classList.add('op0')
                             
-                            console.log(TextData[partKey]);//選択されたシナリオ
+        //                     console.log(TextData[partKey]);//選択されたシナリオ
 
-                            gameState.textEventId++;
-                            gameState.nowPart = partKey
-                            TextPlayer = new ScenarioPlayer(TextData[partKey], gameState)//プレイヤー生成
+        //                     gameState.textEventId++;
+        //                     gameState.nowPart = partKey
+        //                     TextPlayer = new ScenarioPlayer(TextData[partKey], gameState)//プレイヤー生成
                             
-                            // TextPlayerList[gameState.textEventId]=new ScenarioPlayer(TextData[partKey],gameState)//プレイヤー生成
-                            // console.log(gameState);
-                            // console.log(TextPlayerList);
+        //                     // TextPlayerList[gameState.textEventId]=new ScenarioPlayer(TextData[partKey],gameState)//プレイヤー生成
+        //                     // console.log(gameState);
+        //                     // console.log(TextPlayerList);
                             
-                            /*------------------
-                                テキストの処理 
-                            ------------------*/
+        //                     /*------------------
+        //                         テキストの処理 
+        //                     ------------------*/
                             
-                        }
+        //                 }
                     
 
-                    }
+        //             }
     
-                    yesBtn.addEventListener('click',selectEve,false)
-                    noBtn.addEventListener('click',selectEve,false)
-                })
+        //             yesBtn.addEventListener('click',selectEve,false)
+        //             noBtn.addEventListener('click',selectEve,false)
+        //         })
 
-                // スタイル付与
-                const styles = item.style
-                for (const prop in styles) {
-                    if (Object.hasOwnProperty.call(styles, prop)) {
-                        const val = styles[prop]
-                        itemEle.style[prop] = val
-                    }
-                }
+        //         // スタイル付与
+        //         const styles = item.style
+        //         for (const prop in styles) {
+        //             if (Object.hasOwnProperty.call(styles, prop)) {
+        //                 const val = styles[prop]
+        //                 itemEle.style[prop] = val
+        //             }
+        //         }
 
-                // フラグメントにappend
-                eleFragment.appendChild(itemEle)
-            }
-        }
-        // DOMにappend
-        document.getElementById('mapItems').appendChild(eleFragment)
+        //         // フラグメントにappend
+        //         eleFragment.appendChild(itemEle)
+        //     }
+        // }
+        // // DOMにappend
+        // document.getElementById('mapItems').appendChild(eleFragment)
 
         // マップ画面遷移系
         const spb = document.getElementById('speechBubble')
